add working google login/register
