+++ server/server.js
@@ -26,6 +26,11 @@ const paymentMethods = require("./routes/paymentMethods");
 const partners = require("./routes/partners");
 const mainRoute = require("./routes/index");
 
+////////////////////////////////////////////////////////////////////////
+const loginGoogle = require("./routes/loginWithGoogle");
+
+////////////////////////////////////////////////////////////////////////
+
 const app = express();
 
 app.use(cors());
@@ -69,6 +74,10 @@ app.use("/api/shipping-methods", shippingMethods);
 app.use("/api/payment-methods", paymentMethods);
 app.use("/api/partners", partners);
 app.use("/", mainRoute);
+
+////////////////////////////////////////////////////////////////////////
+app.use("/api/logingoogle", loginGoogle);
+////////////////////////////////////////////////////////////////////////
 
 // Server static assets if in production
 if (process.env.NODE_ENV === "production") {
